refactor: split oversized approval and session helpers
